Added new files in order to separate the different code (urls, exception, utils)
